--- back/src/index.spec.ts
@@ -1,7 +1,0 @@
-import {} from './index';
-
-describe('index', () => {
-  test('should be true', () => {
-    expect(1 + 1).toBe(2);
-  });
-});
